requirements.txt
torch>=1.8.0
torchvision>=0.9.0
timm>=0.4.12
numpy>=1.19.2
pandas>=1.2.4
Pillow>=8.2.0
scikit-learn>=0.24.2
tqdm>=4.60.0
matplotlib>=3.4.2